requirements.txt
flask
matplotlib
pyserial
